// src/lib/form-schemas/work-book/confirm-activity.schema.ts
import { z } from "zod"

import { fileSchema } from "../document-management/file.schema"

export const confirmActivitySchema = z.object({
	activityStartTime: z.string(),
	activityEndTime: z.string(),
	executionDate: z.date(),
	comments: z.string().optional(),
	files: z.array(fileSchema).optional(),
})

export type ConfirmActivitySchema = z.infer<typeof confirmActivitySchema>